Small fixes
One being the reputation that you are set at with the pirates to allow landing on all pirate planets. The other makes pirates and bounty hunters hate each other. It at least makes partial sense that bounty hunters are targetting other infamous pirates, and that pirates hate them. It's also quality of life for gameplay.